♻️ refactor Group의 Query 메서드들에 대한 refactoring 진행
♻️ refactor Group의 Query 메서드들에 대한 refactoring 진행
